backend: libs: autopilot: src: mavlink: Dont timeout on message receiving
It can look a bit weird, but with or without the timeout, the code executes almost the same work.

(1) The timeout was added there only to allow that debug message, which I wanted to remove.
(2) As a consequence, the timeout ended up awaiting and cancelling the `mavlink::AsyncMavConnection::recv` call multiple times, potentially unhelpfully in the middle of the reconnection loop.